pyproject.toml
[tool.poetry]
name = "podcastfy"
version = "0.4.1"
description = "An Open Source alternative to NotebookLM's podcast feature: Transforming Multimodal Content into Captivating Multilingual Audio Conversations with GenAI"
authors = ["Tharsis T. P. Souza"]
license = "Apache-2.0"
readme = "README.md"
include = [
    "podcastfy/config.yaml",
    "podcastfy/conversation_config.yaml"
]

[tool.poetry.dependencies]
python = "^3.11"
setuptools = "^75.1.0"
wheel = "^0.44.0"
cython = "^3.0.11"
typer = "^0.12.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
openai = "^1.51.0"
elevenlabs = "^1.9.0"
PyMuPDF = "^1.24.11"
pyyaml = "^6.0.2"
youtube-transcript-api = "^0.6.2"
pydub = "^0.25.1"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.26.0"
pandas = "^2.2.3"
google-generativeai = "^0.8.2"
numpy = ">=1,<2"
langchain = "^0.3.3"
langchain-google-vertexai = "^2.0.4"
langchain-google-genai = "^2.0.1"
httpx = "^0.27.2"
pandoc = "^2.4"
sphinx-rtd-theme = "^3.0.1"
sphinx-autodoc-typehints = "^2.5.0"
nbsphinx = "^0.9.5"
edge-tts = "^6.1.12"
types-pyyaml = "^6.0.12.20240917"
nest-asyncio = "^1.6.0"
ffmpeg = "^1.4"
pytest = "^8.3.3"
pytest-xdist = "^3.6.1"
google-cloud-texttospeech = "^2.21.0"
litellm = "^1.52.0"
langchain-community = "^0.3.5"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
aiohttp = "^3.12.15"
pydantic = "^2.11.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
sphinx = ">=8.0.2"
nbsphinx = "0.9.5"
ipykernel = "^6.29.5"
ffmpeg = "^1.4"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
build_docs = "build_docs:main"



